修复小文件合并理解差异......
