+++ Service/src/main/java/com/hzgc/service/staticrepo/ObjectInfoHandlerTool.java
@@ -10,6 +10,8 @@ import java.io.IOException;
 import java.sql.*;
 import java.sql.Connection;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 
 class ObjectInfoHandlerTool {
@@ -223,6 +225,76 @@ class ObjectInfoHandlerTool {
             objectinfo.put(putOfTNums);
         } catch (IOException e) {
             e.printStackTrace();
+        }
+    }
+
+    /**
+     * 对结果进行排序
+     * @param personObjects 最终返回的一个人员列表
+     * @param staticSortParams 排序参数
+     */
+    public void sortPersonObject(List<PersonObject> personObjects, List<StaticSortParam> staticSortParams) {
+        if (staticSortParams != null) {
+            if (staticSortParams.contains(StaticSortParam.RELATEDASC)) {
+                Collections.sort(personObjects, new Comparator<PersonObject>() {
+                    @Override
+                    public int compare(PersonObject o1, PersonObject o2) {
+                        float sim1 = o1.getSim();
+                        float sim2 = o2.getSim();
+                        return Float.compare(sim1, sim2);
+                    }
+                });
+            }
+            if (staticSortParams.contains(StaticSortParam.RELATEDDESC)) {
+                Collections.sort(personObjects, new Comparator<PersonObject>() {
+                    @Override
+                    public int compare(PersonObject o1, PersonObject o2) {
+                        float sim1 = o1.getSim();
+                        float sim2 = o2.getSim();
+                        return Float.compare(sim2, sim1);
+                    }
+                });
+            }
+            if (staticSortParams.contains(StaticSortParam.IMPORTANTASC)) {
+                Collections.sort(personObjects, new Comparator<PersonObject>() {
+                    @Override
+                    public int compare(PersonObject o1, PersonObject o2) {
+                        int important1 = o1.getImportant();
+                        int important2 = o2.getImportant();
+                        return Integer.compare(important1, important2);
+                    }
+                });
+            }
+            if (staticSortParams.contains(StaticSortParam.IMPORTANTDESC)) {
+                Collections.sort(personObjects, new Comparator<PersonObject>() {
+                    @Override
+                    public int compare(PersonObject o1, PersonObject o2) {
+                        int important1 = o1.getImportant();
+                        int important2 = o2.getImportant();
+                        return Integer.compare(important2, important1);
+                    }
+                });
+            }
+            if (staticSortParams.contains(StaticSortParam.TIMEASC)) {
+                Collections.sort(personObjects, new Comparator<PersonObject>() {
+                    @Override
+                    public int compare(PersonObject o1, PersonObject o2) {
+                        java.sql.Timestamp timestamp1 = o1.getCreatetime();
+                        java.sql.Timestamp timestamp2 = o2.getCreatetime();
+                        return Long.compare(timestamp1.getTime(), timestamp2.getTime());
+                    }
+                });
+            }
+            if (staticSortParams.contains(StaticSortParam.TIMEDESC)) {
+                Collections.sort(personObjects, new Comparator<PersonObject>() {
+                    @Override
+                    public int compare(PersonObject o1, PersonObject o2) {
+                        java.sql.Timestamp timestamp1 = o1.getCreatetime();
+                        java.sql.Timestamp timestamp2 = o2.getCreatetime();
+                        return Long.compare(timestamp2.getTime(), timestamp1.getTime());
+                    }
+                });
+            }
         }
     }
